:art: move some common methods into a utils file
